requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.0,<4.0
requests>=2.25,<3.0
